requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
